+++ src/Custom/OpenAIClient.cs
@@ -148,6 +148,7 @@ public partial class OpenAIClient
     /// the same configuration details.
     /// </remarks>
     /// <returns> A new <see cref="BatchClient"/>. </returns>
+    [Experimental("OPENAI001")]
     public virtual BatchClient GetBatchClient() => new(_pipeline, _options);
 
     /// <summary>
@@ -192,6 +193,7 @@ public partial class OpenAIClient
     /// the same configuration details.
     /// </remarks>
     /// <returns> A new <see cref="FineTuningClient"/>. </returns>
+    [Experimental("OPENAI001")]
     public virtual FineTuningClient GetFineTuningClient() => new(_pipeline, _options);
 
     /// <summary>